hotel_booking/pyproject.toml
[tool.poetry]
package-mode = false
name = "hotel-booking"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.5"
psycopg = {extras = ["binary"], version = "^3.2.9"}
djangorestframework = "^3.16.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.7"
pre-commit = "^4.2.0"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-factoryboy = "^2.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 89
target-version = "py311"
exclude = [".venv", "migrations", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "DJ"]
ignore = ["B008"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
django_find_project = false
pythonpath = ["src"]
python_files = ["tests.py", "test_*.py"]